feat: Task 1 changes (#58)
* feat: Task 1 changes

* feat: combined Dataset to db

* feat: Combined data check added

* feat: requested changes

* feat: Task 1 changes

* feat: combined Dataset to db

* feat: Combined data check added

* feat: requested changes

* feat: requested changes

+++ all_types/myapi_dtypes.py
@@ -101,7 +101,7 @@ class ReqGeodata(Coordinate):
 
 
 class ReqLocation(Coordinate):
-    radius: int
+    radius: float
     bounding_box: list[float]
     page_token: Optional[str] = ""
     text_search: Optional[str] = ""
